+++ src/debugmenu.c
@@ -3,16 +3,21 @@
 
 #include "firedays.h"
 
-void fd_debugmenu_init (fdDebugMenu* m, fdGameState* g) {
-	//TODO
+void fd_debugmenu_init (fdDebugMenu* m, fdScreen* screen, fdCore* core) {
+	m->core = core;
+	m->screen = screen;
 }
 
-void fd_debugmenu_event (fdDebugMenu* m, fdGameState* g) {
-	//TODO
+void fd_debugmenu_event (fdDebugMenu* m) {
+	if (m->mode) {
+		//TODO
+	}
 }
 
-void fd_debugmenu_draw (fdDebugMenu* m, fdGameState* g) {
-	//TODO
+void fd_debugmenu_draw (fdDebugMenu* m) {
+	if (m->mode) {
+		//TODO
+	}
 }