+++ src/filescreen.c
@@ -3,15 +3,16 @@
 
 #include "firedays.h"
 
-void fd_filescreen_init (fdFileScreen* s, fdGameState* g) {
+void fd_filescreen_init (fdFileScreen* s, fdScreen* screen, fdCore* core) {
+	s->core = core;
+	s->screen = screen;
+}
+
+void fd_filescreen_event (fdFileScreen* s) {
 	//TODO
 }
 
-void fd_filescreen_event (fdFileScreen* s, fdGameState* g) {
-	//TODO
-}
-
-void fd_filescreen_draw (fdFileScreen* s, fdGameState* g) {
+void fd_filescreen_draw (fdFileScreen* s) {
 	//TODO
 }