+++ src/pausemenu.c
@@ -3,16 +3,21 @@
 
 #include "firedays.h"
 
-void fd_pausemenu_init (fdPauseMenu* m, fdGameState* g) {
-	//TODO
+void fd_pausemenu_init (fdPauseMenu* m, fdScreen* screen, fdCore* core) {
+	m->core = core;
+	m->screen = screen;
 }
 
-void fd_pausemenu_event (fdPauseMenu* m, fdGameState* g) {
-	//TODO
+void fd_pausemenu_event (fdPauseMenu* m) {
+	if (m->mode) {
+		//TODO
+	}
 }
 
-void fd_pausemenu_draw (fdPauseMenu* m, fdGameState* g) {
-	//TODO
+void fd_pausemenu_draw (fdPauseMenu* m) {
+	if (m->mode) {
+		//TODO
+	}
 }