+++ src/titlescreen.c
@@ -3,15 +3,16 @@
 
 #include "firedays.h"
 
-void fd_titlescreen_init (fdTitleScreen* s, fdGameState* g) {
+void fd_titlescreen_init (fdTitleScreen* s, fdScreen* screen, fdCore* core) {
+	s->core = core;
+	s->screen = screen;
+}
+
+void fd_titlescreen_event (fdTitleScreen* s) {
 	//TODO
 }
 
-void fd_titlescreen_event (fdTitleScreen* s, fdGameState* g) {
-	//TODO
-}
-
-void fd_titlescreen_draw (fdTitleScreen* s, fdGameState* g) {
+void fd_titlescreen_draw (fdTitleScreen* s) {
 	//TODO
 }